Implement runAway() and engageEnemy()
Implemented 2 of the 3 actions needed for Explorer.

George, wasn't sure what int id was referring to, set it to 0 for now,
please change as appropriate. Also, I think there are some things that
could do with abstracting as separate functions for readability, will
speak to you about it tomorrow.

Also, when you set direction movement, I THINK you had Left and Right the
wrong way round (see diff for this commit), so I changed it. If, when you
guys are testing it, it turns out that I was wrong, you will also need to
make the same change to AI.directionToMove().

Connor, look through your comments, some of them still refer to Snails.